+++ src/templates/admin/settings-advanced-software-updates.tsx
@@ -1,0 +1,12 @@
+/**
+ * Software Updates link for advanced settings
+ */
+
+export const SoftwareUpdatesSection = (): JSX.Element => (
+  <div class="stack-md column">
+    <h2>Software Updates</h2>
+    <p>
+      <a href="/admin/update">Check for and install software updates</a>
+    </p>
+  </div>
+);
